+++ cli/docs/changelog.md
@@ -9,7 +9,7 @@
 * Added `PromptDisplayer` to display CLI prompts
 * Added `user metrics consent prompt` to request user consents to collecting and sending metrics
 * Added `override Kurtosis config confirmation prompt` to request user for confirmation when they're trying to initialize the config but it's already created
-* Add `enclave dump` subcommand to dump all the logs for an enclave
+* Add `enclave dump` subcommand to dump all the logs & container specs for an enclave
 
 ### Fixes
 * Limit the max number of Typescript tests running at once to 4, to not overwhelm Docker